# Limit data for absences dashboard