--- a/spec/lib/dashboard_queries_spec.rb
+++ b/spec/lib/dashboard_queries_spec.rb
@@ -5,9 +5,41 @@ RSpec.describe DashboardQueries do
     let!(:pals) { TestPals.create! }
     let!(:student) { FactoryBot.create(:student, school: pals.healey)}
 
+    def create_recent_absence!(student_id)
+      Absence.create!({
+        student_id: student_id,
+        occurred_at: Time.now - 4.days,
+        excused: false,
+        dismissed: false
+      })
+    end
+
     it '#absence_dashboard_data shape of data' do
       json = DashboardQueries.new(pals.healey_laura_principal).absence_dashboard_data(pals.healey)
       expect(json.keys.map(&:to_sym)).to contain_exactly(:students_with_events, :school)
+    end
+
+    it '#absence_dashboard_data has expected fields' do
+      create_recent_absence!(pals.healey.students.first.id)
+      json = DashboardQueries.new(pals.healey_laura_principal).absence_dashboard_data(pals.healey)
+      expect(json[:students_with_events].size).to eq pals.healey.students.size
+      expect(json[:students_with_events].first.keys.map(&:to_sym)).to contain_exactly(*[
+        :absences,
+        :first_name,
+        :grade,
+        :homeroom_label,
+        :id,
+        :last_name,
+        :latest_note
+      ])
+
+      absences_in_response_json = json[:students_with_events].map {|student| student[:absences] }
+      expect(absences_in_response_json.flatten.first.keys.map(&:to_sym)).to contain_exactly(*[
+        :occurred_at,
+        :student_id,
+        :dismissed,
+        :excused
+      ])
     end
 
     it '#tardies_dashboard_data shape of data' do
@@ -18,19 +50,6 @@ RSpec.describe DashboardQueries do
     it '#discipline_dashboard_data shape of data' do
       json = DashboardQueries.new(pals.healey_laura_principal).discipline_dashboard_data(pals.healey)
       expect(json.keys.map(&:to_sym)).to contain_exactly(:students_with_events, :school)
-    end
-
-    it '#individual_student_absence_data has expected fields' do
-      json = DashboardQueries.new(pals.healey_laura_principal).send(:individual_student_absence_data, student)
-      expect(json.keys.map(&:to_sym)).to contain_exactly(*[
-        :absences,
-        :first_name,
-        :grade,
-        :id,
-        :last_name,
-        :homeroom_label,
-        :latest_note
-      ])
     end
 
     it '#individual_student_tardies_data has expected fields' do