# Who is this PR for?
admin

# What problem does this PR fix?
Initially this was an attempt at reducing the date range for the query, but we also want to speed this up for larger schools.

# What does this PR do?
Reworks the querying so there are two database queries, and that these are minimal for the data needed (eg, no old data that will be filtered out in JS, remove older `event_notes`).

# Checklists
+ [x] Author included specs for new code