@kevinrobinson 2018-08-29
This could use a `PerfTest` to measure the impact on the larger New Bedford schools or SHS before shipping.  cc @alexsoble if this is helpful in the short term, feel free to grab this and do the perf test or to just ship it and see, otherwise I can finish this later this week.