@kevinrobinson 2018-09-10
Here's some data on this.  The results are pretty variable, so here's I've included the values that show the most conservative impact:

### 
```
Somerville, using SHS as a test case, and running the PerfTest across the first 10 districtwide users:
  master:        median: 12075ms   p95: 13628ms
  this PR:       median:  4397ms   p95:  5379ms

New Bedford, using Normandin Middle as a test case, and running the PerfTest across the first 5 districtwide users:
  master:        median: 21777ms   p95: 22605ms
  this PR:       median:  7857ms   p95: 11016ms
```

Other test runs were faster/slower, but either way this should help significantly.  It doesn't impact the bytes over the wire, and verified this by looking at the byte size before and after this diff on both districts.