@kevinrobinson 2018-09-10
Confirmed this loads in under 5 seconds for SHS and Normandin in production:

<img width="780" alt="screen shot 2018-09-10 at 1 46 09 pm" src="https://user-images.githubusercontent.com/1056957/45314519-f5e6bd80-b4ff-11e8-84bf-039cdcf2e376.png">

<img width="779" alt="screen shot 2018-09-10 at 1 46 21 pm" src="https://user-images.githubusercontent.com/1056957/45314518-f5e6bd80-b4ff-11e8-94b6-54f0e07a859a.png">